refactor(notifier): extracts styled dialog
